requirements.txt
openai
python-telegram-bot
pdfminer.six
python-docx
openpyxl
xlrd
textract==1.6.3
